Move installer and output endpoints to api-workspace module
