+++ infrastructures/docker/src/main/java/org/eclipse/che/workspace/infrastructure/docker/DockerInfraModule.java
@@ -28,10 +28,8 @@ import org.eclipse.che.workspace.infrastructure.docker.config.env.ProjectsRootEn
 import org.eclipse.che.workspace.infrastructure.docker.config.proxy.DockerProxyModule;
 import org.eclipse.che.workspace.infrastructure.docker.config.volume.ExtraVolumeModule;
 import org.eclipse.che.workspace.infrastructure.docker.environment.DockerEnvironmentTypeModule;
-import org.eclipse.che.workspace.infrastructure.docker.output.OutputService;
 import org.eclipse.che.workspace.infrastructure.docker.server.ServerCheckerFactory;
 import org.eclipse.che.workspace.infrastructure.docker.server.ServerCheckerFactoryImpl;
-import org.eclipse.che.workspace.infrastructure.docker.service.InstallerService;
 import org.eclipse.che.workspace.infrastructure.docker.strategy.ServerEvaluationStrategyModule;
 
 import java.util.Set;
@@ -94,9 +92,6 @@ public class DockerInfraModule extends AbstractModule {
         devMachineEnvVars.addBinding().toProvider(ProjectsRootEnvVariableProvider.class);
         devMachineEnvVars.addBinding().toProvider(JavaOptsEnvVariableProvider.class);
         allMachinesEnvVars.addBinding().toProvider(ApiEndpointEnvVariableProvider.class);
-
-        bind(OutputService.class);
-        bind(InstallerService.class);
 
         install(new DnsResolversModule());
         install(new DockerProxyModule());
