+++ wsmaster/che-core-api-workspace/src/main/java/org/eclipse/che/api/workspace/server/OutputEndpoint.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   Codenvy, S.A. - initial API and implementation
  *******************************************************************************/
-package org.eclipse.che.workspace.infrastructure.docker.output;
+package org.eclipse.che.api.workspace.server;
 
 import org.eclipse.che.api.core.websocket.commons.WebSocketMessageReceiver;
 import org.eclipse.che.api.core.websocket.impl.BasicWebSocketEndpoint;
