+++ wsmaster/che-core-api-workspace/src/main/java/org/eclipse/che/api/workspace/server/OutputService.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   Codenvy, S.A. - initial API and implementation
  *******************************************************************************/
-package org.eclipse.che.workspace.infrastructure.docker.output;
+package org.eclipse.che.api.workspace.server;
 
 import org.eclipse.che.api.core.jsonrpc.commons.RequestHandlerConfigurator;
 import org.eclipse.che.api.core.notification.EventService;
